+++ webpack-extension.config.js
@@ -36,5 +36,6 @@ module.exports = {
         ]
       }
     ]
-  }
+  },
+  stats: 'errors-only',
 };