+++ webpack.config.js
@@ -1,7 +1,6 @@
 'use strict'
 
 const path = require('path');
-const copy = require("copy-webpack-plugin");
 const miniCss = require('mini-css-extract-plugin');
 
 
@@ -20,7 +19,6 @@ module.exports = {
   plugins: [new miniCss()],
   module: {
     rules: [
-      // all files with a `.ts` or `.tsx` extension will be handled by `ts-loader`
       { test: /\.tsx?$/, loader: 'ts-loader' },
       {test: /\.css$/i, use: [miniCss.loader, 'css-loader']}
     ]