+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## 1.7.0
+
+-   Revision History shows current board state
+-   Changed Revision history icon
+-   You can change which file to save to (if you choose to save to a file)
+    -   This is done with a new extension setting, `kanban.saveFiles.pathPreferences`. Usage is detailed in the [readme.](README.md)
+-   Added a new settings panel!
+    -   "Autosave" and "Save to File" settings are now toggle switches in the panel
+    -   Can open global extension settings from the panel
+-   Tasks now inherit the colors of their column
+-   Settings changes are recorded in revision history
+
+-   Bugfixes:
+    -   Column title edits don't show up in history if text wasn't changed
+    -   Default context menu now appears on tasks
+
 ### 1.6.2
 
 -   Internal Changes:
@@ -17,10 +33,10 @@
 
 ### 1.6.1
 
--   Bugfixes
+-   Bugfixes:
     -   Board saves on `Ctrl` + `S` again
     -   Coverage directory no longer included in extension
--   `Ctrl` + `S` shows the "Board Saved" toast
+    -   `Ctrl` + `S` shows the "Board Saved" toast
 
 ## 1.6.0